Added 8 images, green borders, top-bar title and its description. Added Week3Class to show class materials' practice.
